Fixed plex reload, and added a configuration section, which can configure plex automatically
Two new cli groups were added: media and configure. The download method was moved to the media group, and the new command to configure plex was added to the configure group.
The new command connects using the user plex's username and password though the plexapi module, but only saves the token (plus the sections, base url, ..).
Finally, a save_config method was added, now only used in the plex configure method.