Added partial support for kissanime
I rewrote the kissanime extractor entirely, and then limited the thread pool executor rate to 4 episodes worker at a time. It seems to prevent the 'are you human' catpcha from popping and thus block the scraping. Even with 4 instead of 16 threads, the process is relatively quick: between 10 and 15 seconds from the few tests I ran.